- Added post-commit hook to example LangGraph app. - LangGraph app should be working now. - Fixed issues with agentc_testing from before (it seems to work now??).

+++ libs/agentc_testing/agentc_testing/catalog.py
@@ -249,8 +249,10 @@ def environment_factory() -> typing.Callable[..., Environment]:
 
     finally:
         # Clean up the environment.
-        repository_instance.pop().close()
-        shutil.rmtree(directory_instance.pop(), ignore_errors=True)
+        if repository_instance:
+            repository_instance.pop().close()
+        if directory_instance:
+            shutil.rmtree(directory_instance.pop(), ignore_errors=True)
 
 
 if __name__ == "__main__":
